move general-purpose resource fetcher to util
